refactor(usecase): rename generate_preview to generate_curriculum_preview
